+++ next-docs/public/examples/header/Icons.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Header } from '@heathmont/moon-core';
-import { IconProfile } from '@heathmont/moon-assets';
-import { IconAbout } from '@heathmont/moon-assets';
+import { GenericUser, GenericInfo } from '@heathmont/moon-icons';
+import { rem } from "@heathmont/moon-utils";
 
 const Example = () => (
   <div className="flex justify-around items-center w-full">
-    <Header icons={[<IconProfile />, <IconAbout />]}>Title</Header>
+    <Header icons={[<GenericUser fontSize={rem(23)}/>, <GenericInfo fontSize={rem(23)}/>]}>Title</Header>
   </div>
 );