+++ next-docs/public/examples/tabs/IsTop.tsx
@@ -1,10 +1,11 @@
 import React from 'react';
 import { Tabs, TabLink } from '@heathmont/moon-core';
 
-const Example = () => (
+const Example = ({ isVertical }: { isVertical?: boolean }) => (
   <Tabs
     id="nav-test-id-1"
     isTop
+    isVertical={isVertical}
     items={[
       <TabLink href="#2">Featured</TabLink>,
       <TabLink href="#3">Today</TabLink>,