+++ packages/core/src/tabs/styles/TabNav.ts
@@ -1,13 +1,29 @@
 import styled from 'styled-components';
 
-const TabNav = styled.nav<{ isContainer?: boolean }>(({ isContainer }) => [
+const TabNav = styled.nav<{
+  isContainer?: boolean,
+  isVertical?: boolean
+  isTop?: boolean
+  isSegmented?: boolean
+}>(({theme, isContainer, isVertical, isTop, isSegmented}) => [
   {
     position: 'relative',
     width: '100%',
+    display: isContainer ? 'flex' : 'block',
+    '& ul': {
+      position: 'relative',
+      '& > div': {
+        zIndex: 1
+      }
+    }
   },
-  isContainer && {
-    display: 'flex',
-  },
+  isVertical && {
+    flexDirection: 'column',
+    '& ul': {
+      flexDirection: 'column',
+      alignItems: isTop ? 'stretch' : 'flex-start',
+    }
+  }
 ]);
 
 export default TabNav;