+++ next-docs/CHANGELOG.md
@@ -1,5 +1,30 @@
 # Change Log
 
+## 8.5.0
+
+### Minor Changes
+
+- xs button, copy code, icons, caroseul arrows fix
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-accordion@8.5.0
+  - @heathmont/moon-assets@8.5.0
+  - @heathmont/moon-charts@8.5.0
+  - @heathmont/moon-components@8.5.0
+  - @heathmont/moon-core@8.5.0
+  - @heathmont/moon-datepicker@8.5.0
+  - @heathmont/moon-draggabletable@8.5.0
+  - @heathmont/moon-icons@8.5.0
+  - @heathmont/moon-pagination@8.5.0
+  - @heathmont/moon-popover@8.5.0
+  - @heathmont/moon-select@8.5.0
+  - @heathmont/moon-sidebar@8.5.0
+  - @heathmont/moon-table@8.5.0
+  - @heathmont/moon-themes@8.5.0
+  - @heathmont/moon-utils@8.5.0
+
 ## 8.4.1
 
 ### Patch Changes